Conditional sub-expressions are now evaluated lazily
The check that skips the evaluation of sub-expressions has been
refactored and partial evaluation for conds is now possible.
All conditional operators will have to behave this way, their
branches will be evaluated lazily only when that branch will be
followed during execution.

Before this fix recursion simply didn't work, the recursive evaluation
of the recoursive call couldn't be stopped and that lead to stack
exhaustion.