Improved variables substitution for function
The naive approach of replacing recursively variables with their values
right before evaluation have been scraped. It was more readable but it's
hard to replace correctly special forms when you need information that
is only available during evaluation.

With the new implementation, eval() and all the operators gain a <locals,
values> pair that can contain the list of local variables used in the
lambda with the list of values for those variables, that will be used by
the evaluateVariable() function, to do the same thing that the old
replace() method did.